[FIX] website_sale_product_attribute_filter_collapse: fix test error
website_folded is a field of the product.attribute.category model that
is added in the website_sale_product_attribute_filter_category module.
website_sale_product_attribute_filter_collapse is an independent module
and gives an error in this field, so the solution is to remove it when
creating an attribute category in the tests.

TT41385