[FIX] website_sale_hide_price: fix key error when editing snippet
When editing the website, a problem was encountered when inserting and
editing the dynamic product snippet. When changing the filter to
"product accessories" or "recently sold products", an error occurred
accessing the "website_hide_price" key if it didn't exist. The solution
involves modifying the access to the "website_hide_price" key, thus
avoiding the mentioned error. With this correction, the proper
functioning of dynamic product snippets is guaranteed.

TT43493