[FIX] website_sale_checkout_skip_payment: Fix wrong confirmation method.
When an order is confirmed and goes to the path ‘/shop/confirmation’,
the driver method involved is ‘shop_payment_confirmation’ from websitesale.
With the migration, the wrong method was involved as it was a portal
method and its path is ‘/payment/confirmation’, which has nothing to
do with the confirmation of an order.

TT50360